// src/app/models/customer.ts
export interface Customer{
    clientId: number;
    clientNum: number;
    attritionFlag: string;
    customerAge: number;
    gender: string;
    dependentCount: number;
    educationLevel: string;
    maritalStatus: string;
    incomeCategory: number;
    cardCategory: string;
    monthsOnBook: number;
    totalRelationshipCount: number;
    monthsInactive: number;
    contactsCount: number;
    creditLimit: number;
    totalRevolvingBal: number;
    avgOpenToBuy: number;
    amountChangeQt: number;
    transAmt: number;
    transCount: number;
    countChangeQt: number;
    utilRatio: number;
}